privileged-api/requirements.txt
Flask~=1.1.2
flask_restx~=0.2.0
sqlalchemy==1.3.13
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.4
uwsgi==2.0.18
sqlalchemy_utils
pyjwt==1.4.2
flask-bcrypt
requests